# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  name: string;
  role: string;
  username: string;
}

export interface Lead {
  id: number;
  clientName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  telegram?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  product?: string | null;
  /** @nullable */
  tariff?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  netProfit?: number | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  income?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  managerId: number;
  /** @nullable */
  managerName?: string | null;
  /** @nullable */
  paymentDate?: string | null;
  /** @nullable */
  paymentType?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface LeadInput {
  clientName: string;
  phone?: string;
  telegram?: string;
  email?: string;
  product?: string;
  tariff?: string;
  price?: number;
  netProfit?: number;
  source?: string;
  income?: string;
  status: string;
  notes?: string;
  managerId: number;
  paymentDate?: string;
  paymentType?: string;
}

export interface LeadUpdate {
  clientName?: string;
  phone?: string;
  telegram?: string;
  email?: string;
  product?: string;
  tariff?: string;
  price?: number;
  netProfit?: number;
  source?: string;
  income?: string;
  status?: string;
  notes?: string;
  managerId?: number;
  paymentDate?: string;
  paymentType?: string;
}

export interface MoveLeadInput {
  status: string;
}

export interface ActivityNote {
  id: number;
  leadId: number;
  content: string;
  /** @nullable */
  authorName?: string | null;
  createdAt: string;
}

export interface ActivityNoteInput {
  content: string;
}

export interface Payment {
  id: number;
  /** @nullable */
  orderNumber?: number | null;
  clientName: string;
  /** @nullable */
  telegram?: string | null;
  tariff: string;
  revenue: number;
  /** @nullable */
  netProfit?: number | null;
  /** @nullable */
  receivable?: number | null;
  /** @nullable */
  paymentMethod?: string | null;
  paymentDate: string;
  managerId: number;
  /** @nullable */
  managerName?: string | null;
  /** @nullable */
  paymentSchedule?: string | null;
  /** @nullable */
  status?: string | null;
  createdAt?: string;
}

export interface PaymentInput {
  orderNumber?: number;
  clientName: string;
  telegram?: string;
  tariff: string;
  revenue: number;
  netProfit?: number;
  receivable?: number;
  paymentMethod?: string;
  paymentDate: string;
  managerId: number;
  paymentSchedule?: string;
  status?: string;
}

export interface PaymentUpdate {
  clientName?: string;
  telegram?: string;
  tariff?: string;
  revenue?: number;
  netProfit?: number;
  receivable?: number;
  paymentMethod?: string;
  paymentDate?: string;
  managerId?: number;
  paymentSchedule?: string;
  status?: string;
}

export interface DashboardStats {
  totalRevenue: number;
  totalNetProfit: number;
  totalDeals: number;
  conversionRate: number;
  averageCheck: number;
  totalLeads?: number;
  minPlan?: number;
  targetPlan?: number;
  planProgress?: number;
}

export interface CashFlowEntry {
  month: string;
  revenue: number;
  netProfit: number;
  deals: number;
}

export interface ManagerStats {
  managerId: number;
  managerName: string;
  revenue: number;
  netProfit: number;
  deals: number;
  minPlan: number;
  targetPlan: number;
  planProgress?: number;
}

export interface Plan {
  id: number;
  managerId: number;
  /** @nullable */
  managerName?: string | null;
  month: string;
  /** @nullable */
  product?: string | null;
  minPlan: number;
  targetPlan: number;
  maxPlan: number;
  createdAt?: string;
}

export interface PlanInput {
  managerId: number;
  month: string;
  product?: string;
  minPlan: number;
  targetPlan: number;
  maxPlan: number;
}

export interface PlanUpdate {
  minPlan?: number;
  targetPlan?: number;
  maxPlan?: number;
  product?: string;
}

export interface Employee {
  id: number;
  name: string;
  role: string;
  username: string;
  /** @nullable */
  salary?: number | null;
  /** @nullable */
  baseBonus?: number | null;
  /** @nullable */
  multiplier?: number | null;
  /** @nullable */
  minPlan?: number | null;
  /** @nullable */
  targetPlan?: number | null;
  /** @nullable */
  maxPlan?: number | null;
  createdAt?: string;
}

export interface EmployeeInput {
  name: string;
  role: string;
  username: string;
  password: string;
  salary?: number;
  baseBonus?: number;
  multiplier?: number;
  minPlan?: number;
  targetPlan?: number;
  maxPlan?: number;
}

export interface EmployeeUpdate {
  name?: string;
  role?: string;
  username?: string;
  password?: string;
  salary?: number;
  baseBonus?: number;
  multiplier?: number;
  minPlan?: number;
  targetPlan?: number;
  maxPlan?: number;
}

export interface WorkspaceStats {
  managerId: number;
  managerName: string;
  openDeals: number;
  todayTasks: number;
  currentBonus: number;
  amountToTarget: number;
  currentMultiplier: number;
  netProfit: number;
  targetPlan: number;
  minPlan: number;
  planProgress?: number;
}

export type ListLeadsParams = {
/**
 * @nullable
 */
managerId?: number | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
tariff?: string | null;
/**
 * @nullable
 */
month?: string | null;
/**
 * @nullable
 */
search?: string | null;
};

export type ListPaymentsParams = {
/**
 * @nullable
 */
managerId?: number | null;
/**
 * @nullable
 */
tariff?: string | null;
/**
 * @nullable
 */
paymentMethod?: string | null;
/**
 * @nullable
 */
month?: string | null;
/**
 * @nullable
 */
status?: string | null;
/**
 * @nullable
 */
search?: string | null;
};

export type GetDashboardStatsParams = {
/**
 * @nullable
 */
month?: string | null;
};

export type GetManagerComparisonParams = {
/**
 * @nullable
 */
month?: string | null;
};

export type ListPlansParams = {
/**
 * @nullable
 */
managerId?: number | null;
/**
 * @nullable
 */
month?: string | null;
};

export type GetMyStatsParams = {
/**
 * @nullable
 */
month?: string | null;
};