# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get executive dashboard KPIs
      parameters:
        - in: query
          name: month
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
  /dashboard/manager-comparison:
    get:
      operationId: getManagerComparison
      tags: [dashboard]
      summary: Manager performance comparison
      parameters:
        - in: query
          name: month
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManagerStats"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthUser:
      type: object
      required: [id, name, role, username]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        username: { type: string }
...
    DashboardStats:
      type: object
      required: [totalRevenue, totalNetProfit, totalDeals, conversionRate, averageCheck]
      properties:
        totalRevenue: { type: number }
        totalNetProfit: { type: number }
        totalDeals: { type: integer }
        conversionRate: { type: number }
        averageCheck: { type: number }
        totalLeads: { type: integer }
        minPlan: { type: number }
        targetPlan: { type: number }
        planProgress: { type: number }
...
    ManagerStats:
      type: object
      required: [managerId, managerName, revenue, netProfit, deals, minPlan, targetPlan]
      properties:
        managerId: { type: integer }
        managerName: { type: string }
        revenue: { type: number }
        netProfit: { type: number }
        deals: { type: integer }
        minPlan: { type: number }
        targetPlan: { type: number }
        planProgress: { type: number }
...
    Employee:
      type: object
      required: [id, name, role, username]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        username: { type: string }
        salary: { type: ["number", "null"] }
        baseBonus: { type: ["number", "null"] }
        multiplier: { type: ["number", "null"] }
        minPlan: { type: ["number", "null"] }
        targetPlan: { type: ["number", "null"] }
        maxPlan: { type: ["number", "null"] }
        createdAt: { type: string }
    EmployeeInput:
      type: object
      required: [name, role, username, password]
      properties:
        name: { type: string }
        role: { type: string }
        username: { type: string }
        password: { type: string }
        salary: { type: number }
        baseBonus: { type: number }
        multiplier: { type: number }
        minPlan: { type: number }
        targetPlan: { type: number }
        maxPlan: { type: number }
    EmployeeUpdate:
      type: object
      properties:
        name: { type: string }
        role: { type: string }
        username: { type: string }
        password: { type: string }
        salary: { type: number }
        baseBonus: { type: number }
        multiplier: { type: number }
        minPlan: { type: number }
        targetPlan: { type: number }
        maxPlan: { type: number }
...