# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  role: string;
  username: string;
  /** @nullable */
  salary?: number | null;
  /** @nullable */
  baseBonus?: number | null;
  /** @nullable */
  multiplier?: number | null;
  /** @nullable */
  minPlan?: number | null;
  /** @nullable */
  targetPlan?: number | null;
  /** @nullable */
  maxPlan?: number | null;
  createdAt?: string;
}