# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  name: string;
  role: string;
  username: string;
  password: string;
  salary?: number;
  baseBonus?: number;
  multiplier?: number;
  minPlan?: number;
  targetPlan?: number;
  maxPlan?: number;
}